fix for deleteCatalog, deleteTask

+++ src/main/java/ru/leti/wise/task/task/logic/DeleteCatalogOperation.java
@@ -3,6 +3,8 @@ package ru.leti.wise.task.task.logic;
 import com.google.protobuf.Empty;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
+import ru.leti.wise.task.task.error.BusinessException;
+import ru.leti.wise.task.task.error.ErrorCode;
 import ru.leti.wise.task.task.repository.CatalogRepository;
 
 import java.util.UUID;
@@ -14,9 +16,11 @@ public class DeleteCatalogOperation {
     private final CatalogRepository catalogRepository;
 
     public Empty activate(UUID uuid) {
-
+        if(catalogRepository.findById(uuid).isEmpty()){
+            throw new BusinessException(ErrorCode.CATALOG_NOT_FOUND);
+        }
         catalogRepository.deleteById(uuid);
-
+        //todo возможно нужно удаление task_catalog, также проверка
         return Empty.newBuilder().build();
     }
 }
